update home dashboard UI

+++ src/components/dashboard/home/Notes/index.tsx
@@ -1,3 +1,4 @@
+import { useState } from "react"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import Link from "next/link"
 import { Checkbox } from "@/components/ui/checkbox"
@@ -6,13 +7,15 @@ import { api } from "@/convex/_generated/api"
 import { useConvexAuth, useMutation, useQuery } from "convex/react"
 import { toast } from "sonner"
 import { CircleHelp } from "lucide-react"
+import { AddNoteDialog } from "../../Dialogs/notes/AddNoteDialog"
 import * as React from "react"
 
-type CardProps = React.ComponentProps<typeof Card>
+type CardProps = React.ComponentProps<typeof Card>;
 
 const NotesCard = ({ className, ...props }: CardProps) => {
   const { isAuthenticated } = useConvexAuth()
-  const notes:any = useQuery(
+  const [deletedNotes, setDeletedNotes] = useState<string[]>([])
+  const notes: any = useQuery(
     api.notes.getNotes,
     !isAuthenticated ? "skip" : undefined
   )
@@ -20,28 +23,56 @@ const NotesCard = ({ className, ...props }: CardProps) => {
   const deleteNote = useMutation(api.notes.deleteNote)
 
   const handleDeleteNote = async (id: any) => {
+    let proceedWithDelete = true
     try {
-      await deleteNote({
-        id: id,
+      toast("Note has been deleted", {
+        action: {
+          label: "Undo",
+          onClick: () => (proceedWithDelete = false),
+        },
       })
-      toast.success("Note deleted!")
+
+      setDeletedNotes([...deletedNotes, id])
+
+      await new Promise((resolve) => setTimeout(resolve, 3300))
+      if (proceedWithDelete) {
+        await deleteNote({
+          id: id,
+        })
+      } else {
+        toast.info("Delete operation cancelled!")
+        // Revert the local deletion from UI
+        setDeletedNotes(deletedNotes.filter(noteId => noteId !== id))
+      }
     } catch (error) {
-      toast.error("Error deleting Note!")
+      if (proceedWithDelete) {
+        toast.error("Error deleting Note!")
+      } else {
+        toast.info("Error cancelling delete operation!")
+        // Revert the local deletion from UI
+        setDeletedNotes(deletedNotes.filter(noteId => noteId !== id))
+      }
     }
   }
-  
+
   return (
-      <Card className={cn("w-full md:w-2/5 mt-5 md:my-0 border-none", className)} {...props}>
-      <CardHeader>
+    <Card
+      className={cn("w-full md:w-2/5 pb-5 md:pb-20 mt-5 md:my-0 border-none", className)}
+      {...props}
+    >
+      <CardHeader className="flex items-center justify-between flex-row">
         <Link href={"/dashboard/notes"}>
-          <CardTitle className='flex items-center justify-start font-semibold ml-2 text-3xl'>Notes</CardTitle>
+          <CardTitle className="flex items-center justify-start font-semibold ml-2 text-3xl">
+            Notes
+          </CardTitle>
         </Link>
+        <AddNoteDialog />
       </CardHeader>
       <CardContent className="grid gap-4 mt-2">
         <div>
           {notes?.length < 1 ? (
-            <div className="text-black w-full flex flex-col items-center justify-center">
-              <CircleHelp size={56} className="mt-10"/>
+            <div className="text-primaryGray w-full flex flex-col items-center justify-center">
+              <CircleHelp size={56} className="mt-10" />
               <p className="mt-2">No notes were found!</p>
               <Link href={"/dashboard/notes"} className="underline mt-3">
                 Click here to create your first note!
@@ -49,22 +80,27 @@ const NotesCard = ({ className, ...props }: CardProps) => {
             </div>
           ) : (
             <>
-              {notes?.map((note:any, index:any) => (
-              <div
-                key={index}
-                className="mb-3 flex"
-              >
-              
-                <div className="space-y-1 flex items-center justify-start">
-                  <Checkbox id="note" className='mr-5' onCheckedChange={() => handleDeleteNote(note._id)}/>
-                  <div className='flex flex-col items-start justify-center'>
-                      <label className="text-sm font-medium leading-none" htmlFor='note'>
-                      {note.text}
-                      </label>
-                  </div>
-                </div>
-              </div>
-            ))}
+              {notes?.map((note: any, index: any) => (
+                !deletedNotes.includes(note._id) && (
+                  <Card key={index} className="mb-3 flex">
+                    <div className="py-3 px-5 flex items-center justify-center">
+                      <Checkbox
+                        id="note"
+                        className="mr-5"
+                        onCheckedChange={() => handleDeleteNote(note._id)}
+                      />
+                      <div className="flex flex-col items-start justify-center">
+                        <label
+                          className="text-sm font-medium leading-none"
+                          htmlFor="note"
+                        >
+                          {note.text}
+                        </label>
+                      </div>
+                    </div>
+                  </Card>
+                )
+              ))}
             </>
           )}
         </div>
